added smal menu
