+++ old_functions.h
@@ -1,0 +1,11 @@
+void set_PWM_Timer();
+void set_PWM();
+void pwmStart(int PWM_CH,int Duty);
+void pwmStop(int PWM_CH);
+void pwmStopAll();
+void U_V_start(int duty);
+void V_U_start(int duty);
+void U_W_start(int duty);
+void W_U_start(int duty);
+void V_W_start(int duty);
+void W_V_start(int duty);
